+++ fastly/__init__.py
@@ -10,7 +10,7 @@
 """
 
 
-__version__ = "4.1.1"
+__version__ = "4.2.0"
 
 # import ApiClient
 from fastly.api_client import ApiClient