+++ fastly/api/logging_logentries_api.py
@@ -546,7 +546,7 @@ class LoggingLogentriesApi(object):
             format (str): A Fastly [log format string](https://docs.fastly.com/en/guides/custom-log-formats).. [optional] if omitted the server will use the default value of "%h %l %u %t "%r" %&gt;s %b"
             format_version (int): The version of the custom logging format used for the configured endpoint. The logging call gets placed by default in `vcl_log` if `format_version` is set to `2` and in `vcl_deliver` if `format_version` is set to `1`. . [optional] if omitted the server will use the default value of 2
             port (int): The port number.. [optional] if omitted the server will use the default value of 20000
-            token (str): Use token based authentication ([https://logentries.com/doc/input-token/](https://logentries.com/doc/input-token/)).. [optional]
+            token (str): Use token based authentication.. [optional]
             use_tls (LoggingUseTls): [optional]
             region (str): The region to which to stream logs.. [optional]
             _return_http_data_only (bool): response data without head status
@@ -893,7 +893,7 @@ class LoggingLogentriesApi(object):
             format (str): A Fastly [log format string](https://docs.fastly.com/en/guides/custom-log-formats).. [optional] if omitted the server will use the default value of "%h %l %u %t "%r" %&gt;s %b"
             format_version (int): The version of the custom logging format used for the configured endpoint. The logging call gets placed by default in `vcl_log` if `format_version` is set to `2` and in `vcl_deliver` if `format_version` is set to `1`. . [optional] if omitted the server will use the default value of 2
             port (int): The port number.. [optional] if omitted the server will use the default value of 20000
-            token (str): Use token based authentication ([https://logentries.com/doc/input-token/](https://logentries.com/doc/input-token/)).. [optional]
+            token (str): Use token based authentication.. [optional]
             use_tls (LoggingUseTls): [optional]
             region (str): The region to which to stream logs.. [optional]
             _return_http_data_only (bool): response data without head status