+++ fastly/api/whole_platform_ddos_historical_api.py
@@ -1,0 +1,184 @@
+"""
+    Fastly API
+
+    Via the Fastly API you can perform any of the operations that are possible within the management console,  including creating services, domains, and backends, configuring rules or uploading your own application code, as well as account operations such as user administration and billing reports. The API is organized into collections of endpoints that allow manipulation of objects related to Fastly services and accounts. For the most accurate and up-to-date API reference content, visit our [Developer Hub](https://developer.fastly.com/reference/api/)   # noqa: E501
+
+    The version of the OpenAPI document: 1.0.0
+    Contact: oss@fastly.com
+"""
+
+
+import re  # noqa: F401
+import sys  # noqa: F401
+
+from fastly.api_client import ApiClient, Endpoint as _Endpoint
+from fastly.model_utils import (  # noqa: F401
+    check_allowed_values,
+    check_validations,
+    date,
+    datetime,
+    file_type,
+    none_type,
+    validate_and_convert_types
+)
+from fastly.model.platform_ddos_response import PlatformDdosResponse
+
+
+class WholePlatformDdosHistoricalApi(object):
+    """NOTE: This class is auto generated.
+    Do not edit the class manually.
+    """
+
+    def __init__(self, api_client=None):
+        if api_client is None:
+            api_client = ApiClient()
+        self.api_client = api_client
+        self.get_platform_ddos_historical_endpoint = _Endpoint(
+            settings={
+                'response_type': (PlatformDdosResponse,),
+                'auth': [
+                    'token'
+                ],
+                'endpoint_path': '/metrics/platform/ddos',
+                'operation_id': 'get_platform_ddos_historical',
+                'http_method': 'GET',
+                'servers': [
+                    {
+                        'url': "https://api.fastly.com",
+                        'description': "No description provided",
+                    },
+                ]
+            },
+            params_map={
+                'all': [
+                    'start',
+                    'end',
+                    'downsample',
+                ],
+                'required': [],
+                'nullable': [
+                ],
+                'enum': [
+                    'downsample',
+                ],
+                'validation': [
+                ]
+            },
+            root_map={
+                'validations': {
+                },
+                'allowed_values': {
+                    ('downsample',): {
+
+                        "HOUR": "hour",
+                        "DAY": "day"
+                    },
+                },
+                'openapi_types': {
+                    'start':
+                        (str,),
+                    'end':
+                        (str,),
+                    'downsample':
+                        (str,),
+                },
+                'attribute_map': {
+                    'start': 'start',
+                    'end': 'end',
+                    'downsample': 'downsample',
+                },
+                'location_map': {
+                    'start': 'query',
+                    'end': 'query',
+                    'downsample': 'query',
+                },
+                'path_params_allow_reserved_map': {
+                },
+                'collection_format_map': {
+                }
+            },
+            headers_map={
+                'accept': [
+                    'application/json'
+                ],
+                'content_type': [],
+            },
+            api_client=api_client
+        )
+
+    def get_platform_ddos_historical(
+        self,
+        **kwargs
+    ):
+        """Get historical DDoS metrics for the entire Fastly platform  # noqa: E501
+
+        Fetches historical DDoS metrics for the entire Fastly platform.  # noqa: E501
+        This method makes a synchronous HTTP request by default. To make an
+        asynchronous HTTP request, please pass async_req=True
+
+        >>> thread = api.get_platform_ddos_historical(async_req=True)
+        >>> result = thread.get()
+
+
+        Keyword Args:
+            start (str): A valid ISO-8601-formatted date and time, or UNIX timestamp, indicating the inclusive start of the query time range. If not provided, a default is chosen based on the provided `downsample` value.. [optional]
+            end (str): A valid ISO-8601-formatted date and time, or UNIX timestamp, indicating the exclusive end of the query time range. If not provided, a default is chosen based on the provided `downsample` value.. [optional]
+            downsample (str): Duration of sample windows.. [optional] if omitted the server will use the default value of "hour"
+            _return_http_data_only (bool): response data without head status
+                code and headers. Default is True.
+            _preload_content (bool): if False, the urllib3.HTTPResponse object
+                will be returned without reading/decoding response data.
+                Default is True.
+            _request_timeout (int/float/tuple): timeout setting for this request. If
+                one number provided, it will be total request timeout. It can also
+                be a pair (tuple) of (connection, read) timeouts.
+                Default is None.
+            _check_input_type (bool): specifies if type checking
+                should be done one the data sent to the server.
+                Default is True.
+            _check_return_type (bool): specifies if type checking
+                should be done one the data received from the server.
+                Default is True.
+            _spec_property_naming (bool): True if the variable names in the input data
+                are serialized names, as specified in the OpenAPI document.
+                False if the variable names in the input data
+                are pythonic names, e.g. snake case (default)
+            _content_type (str/None): force body content-type.
+                Default is None and content-type will be predicted by allowed
+                content-types and body.
+            _host_index (int/None): specifies the index of the server
+                that we want to use.
+                Default is read from the configuration.
+            async_req (bool): execute request asynchronously
+
+        Returns:
+            PlatformDdosResponse
+                If the method is called asynchronously, returns the request
+                thread.
+        """
+        kwargs['async_req'] = kwargs.get(
+            'async_req', False
+        )
+        kwargs['_return_http_data_only'] = kwargs.get(
+            '_return_http_data_only', True
+        )
+        kwargs['_preload_content'] = kwargs.get(
+            '_preload_content', True
+        )
+        kwargs['_request_timeout'] = kwargs.get(
+            '_request_timeout', None
+        )
+        kwargs['_check_input_type'] = kwargs.get(
+            '_check_input_type', True
+        )
+        kwargs['_check_return_type'] = kwargs.get(
+            '_check_return_type', True
+        )
+        kwargs['_spec_property_naming'] = kwargs.get(
+            '_spec_property_naming', False
+        )
+        kwargs['_content_type'] = kwargs.get(
+            '_content_type')
+        kwargs['_host_index'] = kwargs.get('_host_index')
+        return self.get_platform_ddos_historical_endpoint.call_with_http_info(**kwargs)
+